ws to track chat lists

+++ api/src/entity/chat.rs
@@ -9,7 +9,7 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
     pub name: String,
-    pub owner_id: i64,
+    pub owner_id: i32,
     pub created_at: DateTime,
 }
